Async cluster: Move `Request` to separate file.
